Избавился от матрешки в MovieService
