+++ app/Providers/AppServiceProvider.php
@@ -2,6 +2,8 @@
 
 namespace App\Providers;
 
+use App\Services\MovieService\MovieRepositoryInterface;
+use App\Services\MovieService\MovieOmdbRepository;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -11,7 +13,7 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        //
+        $this->app->bind(MovieRepositoryInterface::class, MovieOmdbRepository::class);
     }
 
     /**
