--- app/Services/MovieService/MovieApiClientInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Services\MovieService;
-
-interface MovieApiClientInterface
-{
-    public function sendRequest(string $imdbId);
-}
