+++ app/Services/MovieService/MovieOmdbRepository.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\Services\MovieService;
+
+use \GuzzleHttp\Client;
+
+class MovieOmdbRepository implements MovieRepositoryInterface
+{
+    private Client $client;
+    private string $apiKey = 'a471b1ee';
+    private string $baseUrl = 'http://www.omdbapi.com/';
+
+    public function __construct(Client $client)
+    {
+        $this->client = $client;
+    }
+
+    public function findMovieById(string $imdbId): ?array
+    {
+        $response = $this->client->request('GET', $this->baseUrl, [
+            'query' => [
+                'apikey' => $this->apiKey,
+                'i' => $imdbId,
+            ],
+        ]);
+
+        return json_decode($response->getBody()->getContents(), true);
+    }
+}
