+++ index.php
@@ -3,14 +3,12 @@
 require_once __DIR__ . '/vendor/autoload.php';
 
 use GuzzleHttp\Client;
-use App\Services\MovieService\MovieOmdbApiClient;
-use App\Services\MovieService\MovieRepository;
+use App\Services\MovieService\MovieOmdbRepository;
 use App\Services\MovieService\MovieService;
 
 $client = new Client();
-$apiClient = new MovieOmdbApiClient($client);
-$repository = new MovieRepository($apiClient);
+$repository = new MovieOmdbRepository($client);
 $service = new MovieService($repository);
 
 $movie = $service->getMovie('tt13157618');
-echo $movie["Title"];
+var_dump($movie['Title']);
